copyDirectory, mvDirectory - daml-project-config -> da-hs-base
changelog_begin
changelog_end
