+++ daml-assistant/daml-helper/BUILD.bazel
@@ -38,6 +38,7 @@ da_haskell_library(
     deps = [
         "//daml-assistant:daml-project-config",
         "//language-support/hs/bindings:hs-ledger",
+        "//libs-haskell/da-hs-base",
     ],
 )
