+++ language-support/ts/codegen/tests/BUILD.bazel
@@ -100,5 +100,6 @@ da_haskell_test(
     deps = [
         "//:sdk-version-hs-lib",
         "//libs-haskell/bazel-runfiles",
+        "//libs-haskell/da-hs-base",
     ],
 ) if not is_windows else None
